+++ src/main/java/ua/kpi/diploma/controltestinghub/dto/ActionVariableDto.java
@@ -2,14 +2,11 @@ package ua.kpi.diploma.controltestinghub.dto;
 
 import lombok.Builder;
 import lombok.Data;
-import ua.netcracker.group3.automaticallytesting.model.Variable;
-
-import java.util.List;
 
 @Builder
 @Data
 public class ActionVariableDto {
-    private Long actionId;
+    private Integer actionId;
     private String actionName;
     private String actionDescription;
     private String variableName;
