+++ src/main/java/ua/kpi/diploma/controltestinghub/service/ActionExecutionService.java
@@ -1,11 +1,15 @@
 package ua.kpi.diploma.controltestinghub.service;
 
 
+import ua.kpi.diploma.controltestinghub.dto.ActionExecutionDto;
+import ua.kpi.diploma.controltestinghub.model.ActionExecutionPassedFailed;
+import ua.kpi.diploma.controltestinghub.util.Pageable;
+
 import java.util.List;
 
 public interface ActionExecutionService {
-    List<ActionExecutionDto> getAllActionExecutions(Long testCaseExecutionId, Pageable pageable);
+    List<ActionExecutionDto> getAllActionExecutions(Integer testCaseExecutionId, Pageable pageable);
     List<ActionExecutionPassedFailed> getActionExecutionPassedFailed(String status);
 
-    Integer getQuantityActionsExecutions(Long testCaseExecutionId,String searchName);
+    Integer getQuantityActionsExecutions(Integer testCaseExecutionId,String searchName);
 }
