+++ src/main/java/ua/kpi/diploma/controltestinghub/service/ActionService.java
@@ -1,6 +1,11 @@
 package ua.kpi.diploma.controltestinghub.service;
 
 
+import ua.kpi.diploma.controltestinghub.dto.ActionDtoWithIdNameVoid;
+import ua.kpi.diploma.controltestinghub.dto.ActionVariableDto;
+import ua.kpi.diploma.controltestinghub.model.Action;
+import ua.kpi.diploma.controltestinghub.util.Pageable;
+
 import java.util.List;
 
 public interface ActionService {
@@ -12,7 +17,7 @@ public interface ActionService {
 
     List<ActionDtoWithIdNameVoid> getAllActionsWithIdName();
 
-    long createAction(String name, String description);
+    Integer createAction(String name, String description);
 
     List<ActionVariableDto> getActionVariableById(Long id);
