+++ src/main/java/ua/kpi/diploma/controltestinghub/service/NotificationService.java
@@ -2,5 +2,5 @@ package ua.kpi.diploma.controltestinghub.service;
 
 public interface NotificationService {
 
-    void addNotifications(long testCaseId, long testCaseExecutionId);
+    void addNotifications(Integer testCaseId, Integer testCaseExecutionId);
 }
