+++ src/main/java/ua/kpi/diploma/controltestinghub/service/TestCaseExecService.java
@@ -1,6 +1,11 @@
 package ua.kpi.diploma.controltestinghub.service;
 
 
+import ua.kpi.diploma.controltestinghub.dto.GroupedTestCaseExecutionDto;
+import ua.kpi.diploma.controltestinghub.dto.TestCaseExecutionDto;
+import ua.kpi.diploma.controltestinghub.dto.TestCaseExecutionsCountsByStartDatesDto;
+import ua.kpi.diploma.controltestinghub.model.TestCaseExecution;
+
 import java.util.List;
 
 public interface TestCaseExecService {
@@ -9,11 +14,11 @@ public interface TestCaseExecService {
 
     Integer countTestCaseExecutions(String testCaseName, String projectName, String status);
 
-    List<TestCaseExecutionDto> getAllTestCaseExecutionWithFailedActionNumber(long limit,long offset, String orderBy, String orderByClause, String testCaseName, String projectName, String status);
+    List<TestCaseExecutionDto> getAllTestCaseExecutionWithFailedActionNumber(Integer limit, Integer offset, String orderBy, String orderByClause, String testCaseName, String projectName, String status);
 
-    Long createTestCaseExecution(Long testCaseId, long userId);
+    Long createTestCaseExecution(Integer testCaseId, Integer userId);
 
-    void updateTestCaseExecution(String status, long testCaseExecutionId);
+    void updateTestCaseExecution(String status, Integer testCaseExecutionId);
 
     List<TestCaseExecutionsCountsByStartDatesDto> getExecutionsByDatesForLastDays(Integer daysFromToday);
