+++ src/main/java/ua/kpi/diploma/controltestinghub/service/TestCaseService.java
@@ -1,17 +1,27 @@
 package ua.kpi.diploma.controltestinghub.service;
 
 import org.springframework.transaction.annotation.Transactional;
+import ua.kpi.diploma.controltestinghub.dto.CreateUpdateTestCaseDto;
+import ua.kpi.diploma.controltestinghub.dto.ScenarioStepDto;
+import ua.kpi.diploma.controltestinghub.dto.TestCaseDto;
+import ua.kpi.diploma.controltestinghub.dto.TestCaseWithUserDto;
+import ua.kpi.diploma.controltestinghub.exception.ValidationException;
+import ua.kpi.diploma.controltestinghub.model.TestCaseTopSubscribed;
+import ua.kpi.diploma.controltestinghub.model.TestCaseUpd;
+import ua.kpi.diploma.controltestinghub.util.Pageable;
+
+import java.util.List;
 
 
 public interface TestCaseService {
 
-    void createTestCase(CreateUpdateTestCaseDto createUpdateTestCaseDto, Long userId);
+    void createTestCase(CreateUpdateTestCaseDto createUpdateTestCaseDto, Integer userId);
 
-    List<ScenarioStepDto> getTestScenarioStep(Long testCaseId);
+    List<ScenarioStepDto> getTestScenarioStep(Integer testCaseId);
 
-    TestCaseDto getTestCase(Long testCaseId);
+    TestCaseDto getTestCase(Integer testCaseId);
 
-    List<TestCaseWithUserDto> getTestCasesWithUser(Long projectID, Pageable pageable, String name) throws ValidationException;
+    List<TestCaseWithUserDto> getTestCasesWithUser(Integer projectID, Pageable pageable, String name) throws ValidationException;
 
     List<TestCaseUpd> getAllTestCases();
 
@@ -21,14 +31,14 @@ public interface TestCaseService {
 
     void updateTestCase(CreateUpdateTestCaseDto createUpdateTestCaseDto);
 
-    void addSubscriber(Long testCaseId, Long userId);
+    void addSubscriber(Integer testCaseId, Integer userId);
 
-    Boolean isFollowedByUser(Long testCaseId, Long userId);
+    Boolean isFollowedByUser(Integer testCaseId, Integer userId);
 
-    void removeSubscriber(Long testCaseId, Long userId);
+    void removeSubscriber(Integer testCaseId, Integer userId);
 
-    void archiveTestCase(Long testCaseId);
+    void archiveTestCase(Integer testCaseId);
 
-    void unarchiveTestCase(Long testCaseId);
+    void unarchiveTestCase(Integer testCaseId);
 
 }
